move doc-by-id and all-paginated to properties/experimental
