+++ addon/properties/index.js
@@ -4,13 +4,8 @@ import {
   paginated
 } from './proxy';
 
-import docById from './doc-by-id';
-import allPaginated from './all-paginated';
-
 export {
   first,
   find,
-  paginated,
-  docById,
-  allPaginated // temporary
+  paginated
 };
